feat(memory,context,string,docs,tests): - arena: robust temp markers (block+offset), correct restore + stats; add multi-block regression test - context: add owns_global_arena; free global only when owned; ownership test - string: fix replace capacity for arena-backed strings (capacity=length+1); add regression - tests: make C99-safe (portable strdup), fix sign-compare - docs: update context API + ownership semantics; document string replace behavior; fix headings; remove unsupported theme option; clean warnings

+++ tests/hash_table_test.c
@@ -336,7 +336,7 @@ void test_resize(void) {
     }
     
     /* Verify size */
-    if (myrtx_hash_table_size(table) != num_entries) {
+    if (myrtx_hash_table_size(table) != (size_t)num_entries) {
         myrtx_arena_free(&arena);
         TEST_FAILED("Size incorrect after resize");
     }
